+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+- Extend decoder to optionally query the `NHTSA Vehicle API` to provide make, model,
+  and vehicle type information.
+- Elaborate `dartdoc` comments and documentation
+
 ## 0.0.1+1
 
 - Restrict the `meta` package version to 1.1.6 to match the Flutter SDK limitations.